[proliferate] split workspace file and git components
