+++ desktop/src/components/workspace/git/PublishDialog.tsx
@@ -1,5 +1,7 @@
-import { useId, type ReactNode } from "react";
+import { useId } from "react";
 import type { CurrentPullRequestResponse } from "@anyharness/sdk";
+import { PublishChangedFiles } from "./PublishChangedFiles";
+import { PublishSection } from "./PublishSection";
 import { AutoHideScrollArea } from "@/components/ui/layout/AutoHideScrollArea";
 import { Button } from "@/components/ui/Button";
 import { Input } from "@/components/ui/Input";
@@ -257,89 +259,5 @@ export function PublishDialog({
         </AutoHideScrollArea>
       </div>
     </ModalShell>
-  );
-}
-
-function PublishSection({
-  children,
-  flush = false,
-}: {
-  children: ReactNode;
-  flush?: boolean;
-}) {
-  return (
-    <section className={flush ? "space-y-3" : "space-y-3 border-t border-border/60 pt-4"}>
-      {children}
-    </section>
-  );
-}
-
-function PublishChangedFiles({
-  groups,
-  scroll,
-}: {
-  groups: {
-    staged: PublishFileRow[];
-    partial: PublishFileRow[];
-    unstaged: PublishFileRow[];
-  };
-  scroll: boolean;
-}) {
-  const content = (
-    <div className="space-y-3">
-      <PublishFileSection title="Staged" files={groups.staged} />
-      <PublishFileSection title="Partially staged" files={groups.partial} />
-      <PublishFileSection title="Unstaged" files={groups.unstaged} />
-    </div>
-  );
-
-  if (!scroll) return content;
-
-  return (
-    <AutoHideScrollArea
-      className="max-h-56 min-h-0"
-      viewportClassName="max-h-56 pr-2"
-    >
-      {content}
-    </AutoHideScrollArea>
-  );
-}
-
-interface PublishFileRow {
-  path: string;
-  additions: number;
-  deletions: number;
-}
-
-function PublishFileSection({
-  title,
-  files,
-}: {
-  title: string;
-  files: PublishFileRow[];
-}) {
-  if (files.length === 0) return null;
-
-  return (
-    <section className="space-y-2">
-      <div className="flex items-center justify-between">
-        <p className="text-xs font-medium uppercase tracking-normal text-muted-foreground">{title}</p>
-        <span className="text-xs text-muted-foreground">{files.length}</span>
-      </div>
-      <div className="overflow-hidden rounded-lg border border-border/60">
-        {files.map((file) => (
-          <div
-            key={`${title}:${file.path}`}
-            className="flex items-center gap-3 border-b border-border/60 bg-background px-3 py-2 last:border-b-0"
-          >
-            <span className="min-w-0 flex-1 truncate text-start text-xs text-foreground [direction:rtl]" title={file.path}>
-              <span className="[direction:ltr] [unicode-bidi:plaintext]">{file.path}</span>
-            </span>
-            <span className="shrink-0 text-xs tabular-nums text-git-green">+{file.additions}</span>
-            <span className="shrink-0 text-xs tabular-nums text-git-red">-{file.deletions}</span>
-          </div>
-        ))}
-      </div>
-    </section>
   );
 }
